winui: extract common DialogWindow to shared project
Extract the GitHubDialogWindow into a shared control that can be used by
multiple consumers. Also simplify the IGui interface and optimise
creating a DialogWindow window with a control.
