--- src/windows/GitHub.UI.Windows/Dialog/GitHubDialogWindow.xaml
@@ -1,65 +1,0 @@
-﻿<Window x:Class="GitHub.UI.Dialog.GitHubDialogWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Width="414"
-        Height="460"
-        MinWidth="414"
-        MinHeight="460"
-        Background="{DynamicResource WindowPanelBrush}"
-        Foreground="{DynamicResource WindowPanelTextBrush}"
-        FontFamily="Segoe UI"
-        FontSize="12"
-        FontStretch="Normal"
-        FontStyle="Normal"
-        FontWeight="Normal"
-        SnapsToDevicePixels="True"
-        UseLayoutRounding="True"
-        WindowStartupLocation="CenterOwner"
-        WindowStyle="None"
-        WindowState="Normal"
-        ResizeMode="NoResize"
-        ShowInTaskbar="true"
-        Title="{Binding Title}">
-
-    <Window.Resources>
-        <ResourceDictionary Source="..\Styles.xaml" />
-    </Window.Resources>
-
-    <WindowChrome.WindowChrome>
-        <WindowChrome
-            CaptionHeight="0"
-            ResizeBorderThickness="5" />
-    </WindowChrome.WindowChrome>
-
-    <Border MouseDown="Border_MouseDown"
-            BorderThickness="1"
-            Background="{DynamicResource WindowPanelBrush}"
-            BorderBrush="{DynamicResource WindowBorderBrush}">
-        <DockPanel>
-            <DockPanel DockPanel.Dock="Top">
-                <Label Content="{Binding Title}"
-                        Foreground="{DynamicResource WindowPanelTextBrush}"
-                        DockPanel.Dock="Left"
-                        Margin="9,5,0,0" />
-
-                <!-- close button and padding -->
-                <Button x:Name="CloseButton"
-                        Style="{DynamicResource WindowCloseButton}"
-                        DockPanel.Dock="Right"
-                        HorizontalAlignment="Right"
-                        VerticalAlignment="Top"
-                        IsCancel="True"
-                        Click="CloseButton_Click">
-                    <Path
-                        Width="16"
-                        Height="16"
-                        Fill="{Binding Path=(TextElement.Foreground), RelativeSource={RelativeSource Self}}"
-                        Data="F1M8.583,8L13,12.424 12.424,13 8,8.583 3.576,13 3,12.424 7.417,8 3,3.576 3.576,3 8,7.417 12.424,3 13,3.576z"/>
-                </Button>
-
-            </DockPanel>
-
-            <ContentControl x:Name="ContentHolder"/>
-        </DockPanel>
-    </Border>
-</Window>
