+++ src/windows/GitHub.UI.Windows/Login/Login2FaViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using System.Windows.Input;
 using Microsoft.Git.CredentialManager;
 using Microsoft.Git.CredentialManager.UI;
+using Microsoft.Git.CredentialManager.UI.ViewModels;
 
 namespace GitHub.UI.Login
 {
