+++ src/windows/GitHub.UI.Windows/Login/LoginCredentialsView.xaml
@@ -7,8 +7,6 @@
              xmlns:sharedControls="clr-namespace:Microsoft.Git.CredentialManager.UI.Controls;assembly=Microsoft.Git.CredentialManager.UI"
              xmlns:sharedConverters="clr-namespace:Microsoft.Git.CredentialManager.UI.Converters;assembly=Microsoft.Git.CredentialManager.UI"
              xmlns:octicons="clr-namespace:GitHub.UI.Octicons"
-             Background="{DynamicResource WindowPanelBrush}"
-             Foreground="{DynamicResource WindowPanelTextBrush}"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              mc:Ignorable="d"
              d:DesignWidth="414"
