+++ src/windows/GitHub.UI.Windows/Login/LoginCredentialsViewModel.cs
@@ -1,6 +1,7 @@
 using System.ComponentModel;
 using System.Security;
 using Microsoft.Git.CredentialManager.UI;
+using Microsoft.Git.CredentialManager.UI.ViewModels;
 
 namespace GitHub.UI.Login
 {
