+++ src/windows/Shared.UI.Windows/Controls/DialogWindow.xaml
@@ -1,0 +1,126 @@
+﻿<Window x:Class="Microsoft.Git.CredentialManager.UI.Controls.DialogWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" mc:Ignorable="d"
+        xmlns:viewModels="clr-namespace:Microsoft.Git.CredentialManager.UI.ViewModels"
+        Width="414"
+        Height="460"
+        MinWidth="414"
+        MinHeight="460"
+        Background="{DynamicResource WindowPanelBrush}"
+        Foreground="{DynamicResource WindowPanelTextBrush}"
+        FontFamily="Segoe UI"
+        FontSize="12"
+        FontStretch="Normal"
+        FontStyle="Normal"
+        FontWeight="Normal"
+        SnapsToDevicePixels="True"
+        UseLayoutRounding="True"
+        WindowStartupLocation="CenterOwner"
+        WindowStyle="None"
+        WindowState="Normal"
+        ResizeMode="NoResize"
+        ShowInTaskbar="true"
+        d:DataContext="{d:DesignInstance viewModels:WindowViewModel}"
+        Title="{Binding Title}">
+
+    <Window.Resources>
+        <Color x:Key="WindowPanelColor">#FBFBFB</Color>
+        <Color x:Key="WindowPanelTextColor">#1E1E1E</Color>
+        <Color x:Key="WindowBorderColor">#CCCEDB</Color>
+        <Color x:Key="WindowButtonColor">#FBFBFB</Color>
+        <Color x:Key="WindowButtonGlyphColor">#332C21</Color>
+        <Color x:Key="WindowButtonHoverColor">#FCD8A2</Color>
+        <Color x:Key="WindowButtonHoverGlyphColor">#332C21</Color>
+        <Color x:Key="WindowButtonHoverBorderColor">#FFFFFF</Color>
+        <Color x:Key="WindowButtonDownColor">#FCD8A2</Color>
+        <Color x:Key="WindowButtonDownGlyphColor">#332C21</Color>
+        <Color x:Key="WindowButtonDownBorderColor">#FFFFFF</Color>
+
+        <SolidColorBrush x:Key="WindowPanelBrush" Color="{DynamicResource WindowPanelColor}" />
+        <SolidColorBrush x:Key="WindowPanelTextBrush" Color="{DynamicResource WindowPanelTextColor}" />
+        <SolidColorBrush x:Key="WindowBorderBrush" Color="{DynamicResource WindowBorderColor}" />
+        <SolidColorBrush x:Key="WindowButtonBrush" Color="{DynamicResource WindowButtonColor}" />
+        <SolidColorBrush x:Key="WindowButtonGlyphBrush" Color="{DynamicResource WindowButtonGlyphColor}" />
+        <SolidColorBrush x:Key="WindowButtonHoverBrush" Color="{DynamicResource WindowButtonHoverColor}" />
+        <SolidColorBrush x:Key="WindowButtonHoverGlyphBrush" Color="{DynamicResource WindowButtonHoverGlyphColor}" />
+        <SolidColorBrush x:Key="WindowButtonHoverBorderBrush" Color="{DynamicResource WindowButtonHoverBorderColor}" />
+        <SolidColorBrush x:Key="WindowButtonDownBrush" Color="{DynamicResource WindowButtonDownColor}" />
+        <SolidColorBrush x:Key="WindowButtonDownGlyphBrush" Color="{DynamicResource WindowButtonDownGlyphColor}" />
+        <SolidColorBrush x:Key="WindowButtonDownBorderBrush" Color="{DynamicResource WindowButtonDownBorderColor}" />
+
+        <Style x:Key="WindowCloseButton" TargetType="{x:Type Button}">
+        <Setter Property="FocusVisualStyle" Value="{x:Null}"/>
+        <Setter Property="Background" Value="{DynamicResource WindowButtonBrush}"/>
+        <Setter Property="Foreground" Value="{DynamicResource WindowButtonGlyphBrush}"/>
+        <Setter Property="BorderBrush" Value="Transparent"/>
+        <Setter Property="BorderThickness" Value="0"/>
+        <Setter Property="HorizontalContentAlignment" Value="Center"/>
+        <Setter Property="VerticalContentAlignment" Value="Center"/>
+        <Setter Property="Width" Value="46" />
+        <Setter Property="Height" Value="33" />
+        <Style.Triggers>
+            <Trigger Property="IsDefaulted" Value="true">
+                <Setter Property="Background" Value="{DynamicResource WindowButtonBrush}"/>
+                <Setter Property="Foreground" Value="{DynamicResource WindowButtonGlyphBrush}"/>
+            </Trigger>
+            <Trigger Property="IsKeyboardFocused" Value="true">
+                <Setter Property="Background" Value="{DynamicResource WindowButtonHoverBrush}"/>
+                <Setter Property="Foreground" Value="{DynamicResource WindowButtonHoverGlyphBrush}"/>
+                <Setter Property="BorderBrush" Value="{DynamicResource WindowButtonHoverBorderBrush}"/>
+            </Trigger>
+            <Trigger Property="IsMouseOver" Value="true">
+                <Setter Property="Background" Value="{DynamicResource WindowButtonHoverBrush}"/>
+                <Setter Property="Foreground" Value="{DynamicResource WindowButtonHoverGlyphBrush}"/>
+                <Setter Property="BorderBrush" Value="{DynamicResource WindowButtonHoverBorderBrush}"/>
+            </Trigger>
+            <Trigger Property="IsPressed" Value="true">
+                <Setter Property="Background" Value="{DynamicResource WindowButtonDownBrush}"/>
+                <Setter Property="Foreground" Value="{DynamicResource WindowButtonDownGlyphBrush}"/>
+                <Setter Property="BorderBrush" Value="{DynamicResource WindowButtonDownBorderBrush}"/>
+            </Trigger>
+            <Trigger Property="IsFocused" Value="true">
+                <Setter Property="Background" Value="{DynamicResource WindowButtonHoverBrush}"/>
+                <Setter Property="Foreground" Value="{DynamicResource WindowButtonHoverGlyphBrush}"/>
+                <Setter Property="BorderBrush" Value="{DynamicResource WindowButtonHoverBorderBrush}"/>
+            </Trigger>
+        </Style.Triggers>
+    </Style>
+    </Window.Resources>
+
+    <WindowChrome.WindowChrome>
+        <WindowChrome
+            CaptionHeight="0"
+            ResizeBorderThickness="5" />
+    </WindowChrome.WindowChrome>
+
+    <Border MouseDown="Border_MouseDown"
+            BorderThickness="1"
+            Background="{DynamicResource WindowPanelBrush}"
+            BorderBrush="{DynamicResource WindowBorderBrush}">
+        <DockPanel>
+            <DockPanel DockPanel.Dock="Top">
+                <Label Content="{Binding Title}"
+                        Foreground="{DynamicResource WindowPanelTextBrush}"
+                        DockPanel.Dock="Left"
+                        Margin="9,5,0,0" />
+                <Button Style="{DynamicResource WindowCloseButton}"
+                        DockPanel.Dock="Right"
+                        HorizontalAlignment="Right"
+                        VerticalAlignment="Top"
+                        IsCancel="True"
+                        Click="CloseButton_Click">
+                    <Path
+                        Width="16"
+                        Height="16"
+                        Fill="{Binding Path=(TextElement.Foreground), RelativeSource={RelativeSource Self}}"
+                        Data="F1M8.583,8L13,12.424 12.424,13 8,8.583 3.576,13 3,12.424 7.417,8 3,3.576 3.576,3 8,7.417 12.424,3 13,3.576z"/>
+                </Button>
+
+            </DockPanel>
+
+            <ContentControl x:Name="ContentHolder" Margin="30,30,30,0"/>
+        </DockPanel>
+    </Border>
+</Window>
