+++ src/windows/Shared.UI.Windows/ViewModels/WindowViewModel.cs
@@ -1,7 +1,8 @@
-﻿using System;
-using Microsoft.Git.CredentialManager.UI;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+using System;
 
-namespace GitHub.UI
+namespace Microsoft.Git.CredentialManager.UI.ViewModels
 {
     public abstract class WindowViewModel : ViewModel
     {
